requirements.txt
Flask
Flask-Cors
fpdf
python-dotenv
